refactor: modal에 필요한 style 상수 이동함 style 상수는 따로 모아둠
