+++ client/src/components/Alert/style.js
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
 import {
   MODAL, CREATE_ROOM_BUTTON, TITLE, HEADER,
-} from '../../constants/modal';
+} from '../Style/Modal/constants';
 import { setBorderAndRadius } from '../Style/util';
 import { LOBBY, BUTTON_EFFECT } from '../Style/Lobby/constants';
